% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebpageImageScraper.R
\name{imageScraper}
\alias{imageScraper}
\title{Webpage Image Scraper}
\usage{
imageScraper(url)
}
\arguments{
\item{url}{The URL of the webpage.}
}
\value{
A character vector containing the images of the webpage.
}
\description{
This function scrapes the images of a webpage.
}
