% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebpageLinkScraper.R
\name{linkScraper}
\alias{linkScraper}
\title{Webpage Link Scraper}
\usage{
linkScraper(url)
}
\arguments{
\item{url}{The URL of the webpage.}
}
\value{
A character vector containing the links of the webpage.
}
\description{
This function scrapes the links of a webpage.
}
