% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebpageTableScraper.R
\name{tableScraper}
\alias{tableScraper}
\title{Webpage Table Scraper}
\usage{
tableScraper(url)
}
\arguments{
\item{url}{The URL of the webpage.}
}
\value{
A list containing the tables of the webpage.
}
\description{
This function scrapes the tables of a webpage.
}
