% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebpageTextScraper.R
\name{textScraper}
\alias{textScraper}
\title{Webpage Text Scraper}
\usage{
textScraper(url)
}
\arguments{
\item{url}{The URL of the webpage.}
}
\value{
A character vector containing the text of the webpage.
}
\description{
This function scrapes the text of a webpage.
}
