% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebpageTitleScraper.R
\name{titleScraper}
\alias{titleScraper}
\title{WebpageTitleScraper}
\usage{
titleScraper(url)
}
\arguments{
\item{url}{The URL of the webpage.}
}
\value{
A character vector containing the titles of the webpage.
}
\description{
This function scrapes the titles of a webpage.
}
